Pluralize or singularize to-many key name based on inclusion state
In #13 another bug was noticed, where the to-many relationship,
in case where `included` tree was not present,
did not generate expected key name (it was not inflected,
so whatever was the original relationship_key form, it was
persisting).

This commit includes ActiveSupport::Inflector to handle
pluralizing and singularizing names based on the inclusion state.